initial demo of plugins working at all
This is not a great design, but it's a success at all -- you can
use the generated plugin, if you build it, and it works. For instance,
with a sample table where 'neighbors' is a field with about 0.2
density across 1024 columns:

	CountNeighbors(Row(neighbors=0))

came out as 60, while the raw count was 185. That's at least plausible.

The design is a bit crufty, but I think it's a valid proof of
concept that we can, in principle, make this work.

There's also a Neighbors function to show how straight bitmap ops work.