Various importRoaring performance changes
Some of these affect other use cases, such as anything
using importPositions, potentially.

Don't compute the cache count for a row when not using a cache.

Avoid a pathological case where importRoaring decides to use
importPositions, then goes ahead and does a snapshot anyway.

Allow closing and reopening a file (and thus redirecting the
opWriter to the new file) without re-loading the entire thing
by unmarshalling the bitmap. This should dramatically reduce
the cost of imports that trigger snapshots, which is a lot of
imports. In general, this means we reload only if the bitmap
we wrote isn't our current storage bitmap.

When removing, Containers should uncache the key in question;
otherwise if you clear a row with only one container in it,
the next check of that row still shows the container, because
Get() still returns it even though it was just deleted.

When snapshotting, instead of reloading the storage, just point
f.storage at the bitmap you just wrote. This does mean that
we can end up with mmapped containers pointing at a memory
map that will never be unmapped; we will perhaps want to fix
that later, but we aren't remapping or anything, so it's
a limited cost. (Also means the original fragment file can
be kept from actually being deleted until the process exits
or the fragment is Really For Sure closed, which is less
frequent.)