try to make unmarshal-import faster
RoaringImport works by creating a shiny new bitmap, then possibly
converting it to a slice of values, or possibly performing various
bitmap-level ops on it. This seems expensive. We implement a
streamlined quasi-unmarshal that iterates through and omits parts
of the task in some cases; for instance, there's no need to
do work to unmarshal containers to clear bits if the existing bitmap
hasn't got those bits.