don't clone memory-backed containers during import
It's pretty common to end up just cloning the synthesized
temporary logical container -- but this is actually unnecessary,
as it doesn't otherwis exist. We just need to copy the container
structure and keep using that storage. That does mean that the
containers dumped into the bitmap might hold the importRoaring data
buffer for a while, but if there's a significant number of cloned
containers, they're likely to go away shortly thereafter.

But mark it as mapped, so if anything else changes, we'll
reallocate. But most likely we'll be snapshotting and not care.