fancy tweaks to handle the small cases
When importing into an empty bitmap, we don't need to spend
time doing the fancy put/update/etc stuff; we can just dump
containers in.

When we *are* updating, if we're setting a smallish number
of bits, it's cheaper to determine the set of new bits (bits
to set, minus any existing bits) and work from that, instead
of doing the union and then computing the difference between
the union and the previous state.

This doesn't get performance back to where it was for the
small-N cases, but it chops 30-50% off memory and runtime
for those cases. Which are, after all, tiny.