don't try to make an iterator for no data
Sometimes we don't have any data to remap roaring storage to,
but we still want to unmap all the existing containers.

Possibly that should have been a different function, at the
time it seemed "obvious" that the empty set of data is still a
set of data to remap to. Anyway, don't try to iterate empty
data, just leave the iterator nil and go on.