avoid copying but give us a one-snapshot safety margin
Copying data on row fetch is insanely expensive and breaks
things under heavy load.

Not copying data on row fetch means that if something is still
using a rowcache entry with mmapped storage when we unmap the
storage, that's a segfault.

What if we unmapped things *one snapshot later*, thus, giving
any queries in flight the interval between snapshots to
complete?