move shard distribution logic down to index level  520ms < Thu 17:06:02
This commit moves the shard distribution logic from the cluster level to the index level. This allows multiple indexes to have different shard distributing algorithms, which are all stored in a map in cluster. The shard distributing algorithms are specified through server options, and thus cannot be set at runtime. Shard distributing algorithms for each index are specified through IndexOptions, which are verified for validity.