reopen file when truncating, use int64 offset
Make the offset int64-flavored (this doesn't quite fix the
underlying risk of the offset wrapping, but we are a ways from
that mattering still). Also, when truncating a file, reopen
it. This is because Truncate doesn't change the write position
offset, but you can't Seek on a file opened with O_APPEND.