ensure row cache gets initialized in openStorage
We're getting panics from a nil rowCache in some cases, and
it doesn't seem like this makes sense -- we shouldn't be
able to get out of openStorage without the rowCache being
initialized, without getting an error that prevents startup.

But we did, so let's just be a bit more careful. Also,
because it's a cheap check, we'll initialize the rowCache if
we somehow get to unprotectedRow without it being
initialized. Even though that shouldn't ever happen.